Add filters to weighted ratings
Write the public method getSimilarRatingsByFilter, which is similar to the getSimilarRatings method but has one additional Filter parameter named filterCriteria and uses that filter to access and rate only those movies that match the filter criteria.

This was fairly simple: 1) copy getSimilarRatings  method.  2) change method name and update formal parameter list.  3)  change the one line where it calls the movieDB to use the filter parameter in this DB call rather than making a new TrueFilter.